First "working" version
I think I've stressed the Spotify Api too hard so I must take a break from testing the program. Added the ability to create a new playlist, or add to an existing playlist. This code is functional not pretty, so I will go back and clean up what I can